+++ src/main/java/sixgaezzang/sidepeek/users/domain/Password.java
@@ -1,7 +1,7 @@
 package sixgaezzang.sidepeek.users.domain;
 
-import static sixgaezzang.sidepeek.common.util.ValidationUtils.validateNotBlank;
-import static sixgaezzang.sidepeek.common.util.ValidationUtils.validatePassword;
+import static sixgaezzang.sidepeek.common.util.validation.ValidationUtils.validateNotBlank;
+import static sixgaezzang.sidepeek.common.util.validation.ValidationUtils.validatePassword;
 import static sixgaezzang.sidepeek.users.exception.UserErrorCode.BLANK_PASSWORD;
 import static sixgaezzang.sidepeek.users.exception.UserErrorCode.INVALID_PASSWORD_FORMAT;
